Creation <io> file/Work on the save output function
