+++ multinmrfit/ui.py
@@ -26,7 +26,6 @@ from PIL import Image, ImageTk
 import matplotlib
 import matplotlib.pyplot as plt
 from matplotlib.backends.backend_tkagg import FigureCanvasTkAgg
-from matplotlib.backends.backend_pdf import PdfPages
 
 # Import math libraries
 import pandas as pd
@@ -34,7 +33,6 @@ import numpy as np
 
 # Import our own libraries
 import multinmrfit.run as nmrr
-import multinmrfit.multiplets as nfm
 import multinmrfit.fitting as nff
 
 matplotlib.use("TkAgg")
@@ -229,180 +227,7 @@ def run_user_clustering(figures, colors_plot, peak_picking_threshold, peak_picki
 # Final Plots
 ###################################
 
-def getList(dict):
-    return [k for k in dict.keys()]
 
-def getIntegral(x_fit_, _multiplet_type_, fit_par):
-    d_mapping = nfm.mapping_multiplets()[0]
-    _multiplet_type_function = d_mapping[_multiplet_type_]["f_function"]
-    y = _multiplet_type_function(x_fit_, *fit_par)
-    integral = np.sum(y)*(x_fit_[1]-x_fit_[0])
-    return integral
-
-def single_plot_function(r, x_scale, intensities, fit_results, x_fit, d_id, scaling_factor, analysis_type, output_path, output_folder,output_name,i=None):    
-    fig, ax = plt.subplots(1, 1)
-    fig.set_size_inches([11.7,8.3])
-    ax.plot(
-        x_scale,
-        intensities[r if analysis_type == 'Pseudo2D' else i ,:],
-        color='b',
-        ls='None',
-        marker='o',
-        markersize=7
-        )    
-    ax.invert_xaxis()
-    res = fit_results.loc[r if analysis_type == 'Pseudo2D' else i ].iloc[:].values.tolist()
-
-    sim = nff.simulate_data(
-        x_fit,
-        res,
-        d_id,
-        scaling_factor
-        )
-
-    ax.plot(
-        x_fit, 
-        sim, 
-        'r-', 
-        lw=1,
-        label='fit')
-
-    res_num = r+1 if analysis_type == 'Pseudo2D' else r
-    ax.text(0.05,0.9,"Spectra : " +str(res_num),transform=ax.transAxes,fontsize=20)  
-    ax.set_ylabel('Intensity')
-    ax.set_xlabel(r'$^1H$ $(ppm)$')
-
-    plt.subplots_adjust(
-        left = 0.1,
-        #bottom = 0.04,
-        right = 0.96,
-        top = 0.96,
-        wspace = 0.3,
-        hspace = 0.3,
-    )
-    path_2_save = Path(output_path,output_folder,'plot_ind')
-    path_2_save.mkdir(parents=True,exist_ok=True)
-
-    plt.savefig(str(Path(path_2_save,output_name+'_'+str(res_num)+'.pdf')))
-    plt.close(fig)
-
-def save_output_data(
-    user_input         = 'user_input',
-    fit_results         = 'fit_results', 
-    intensities         = 'intensities',    
-    x_scale             = 'x_scale',
-    spectra_to_fit      =  'spectra_to_fit',
-    Peak_Picking_data   =  None,
-    scaling_factor      =  None
-    ):
-
-    output_path         =   user_input['output_path']
-    output_folder       =   user_input['output_folder']
-    output_name         =   user_input['output_name']
-    analysis_type       =   user_input['analysis_type']
-    data_exp_no         =   user_input['data_exp_no']
-
-    logger.info('Save data to text file ')
-
-    x_fit = np.linspace(np.min(x_scale),np.max(x_scale),2048)
-    d_id = nff.Initial_Values(Peak_Picking_data, x_fit, scaling_factor)[0]
-
-    cluster_list = getList(d_id)
-
-    Fit_results = fit_results.apply(pd.to_numeric)
-    Fit_results_text= Fit_results.round(9)
-
-    Path(output_path,output_folder).mkdir(parents=True,exist_ok=True)
-    
-    d_mapping, _, d_parameters = nfm.mapping_multiplets()
-
-    for i in cluster_list:        
-        #Check ifoutput file exists 
-
-        col = range(d_id[i][1][0],d_id[i][1][1])
-        _multiplet_type_ = d_parameters[len(col)]
-        _multiplet_params_ = d_mapping[_multiplet_type_]['params']
-
-        mutliplet_results = Fit_results_text[Fit_results_text.columns.intersection(col)]
-  
-        mutliplet_results.columns = _multiplet_params_
-
-        mutliplet_results["integral"] = [scaling_factor*getIntegral(x_fit, _multiplet_type_, row.tolist()) for index, row in mutliplet_results.iterrows()]
-        mutliplet_results["Amp"] = scaling_factor*mutliplet_results["Amp"]
-
-        if analysis_type == 'Pseudo2D':
-            mutliplet_results.insert(loc = 0, column = 'exp_no' , value = np.array([data_exp_no]*len(spectra_to_fit)))
-            mutliplet_results.insert(loc = 1, column = 'row_id' , value = spectra_to_fit)
-        elif analysis_type == '1D_Series':
-            mutliplet_results.insert(loc = 0, column = 'exp_no' , value = spectra_to_fit)
-            mutliplet_results.insert(loc = 1, column = 'row_id' , value = [1]*len(spectra_to_fit))
-
-        mutliplet_results.set_index('exp_no', inplace=True)
-        mutliplet_results.to_csv(
-            str(Path(output_path,output_folder,output_name+'_'+str(_multiplet_type_)+'_'+str(i)+'.txt')), 
-            index=True, 
-            sep = '\t'
-            )
-    logger.info('Save data to text file -- Complete')
-
-    logger.info('Save plot to pdf')
-
-    
-    # root, close_button, progress_bars = init_progress_bar_windows(len_progresses = [len(speclist)],title='Output data in pdf',progress_bar_label=[None]) 
-
-    for r in spectra_to_fit:
-        if analysis_type == '1D_Series':
-            i = spectra_to_fit.index(r)
-        single_plot_function(
-                r, 
-                x_scale, 
-                intensities,        
-                fit_results, 
-                x_fit, 
-                d_id, 
-                scaling_factor, 
-                analysis_type,
-                output_path,   
-                output_folder,
-                output_name,
-                i  
-            )
-            
-    # with PdfPages(Path(output_path,output_folder,output_name+'.pdf')) as pdf:   
-    #     matplotlib.pyplot.switch_backend('Agg') 
-    #     for r in range(len(speclist)):
-    #         single_plot_function(
-    #             r, 
-    #             pdf, 
-    #             x_scale, 
-    #             intensities,        
-    #             fit_results, 
-    #             x_fit, 
-    #             d_id, 
-    #             scaling_factor, 
-    #             id_spectra, 
-    #             speclist
-    #         )    
-
-        # threads = []
-        # threads.append(MyApp_Plotting(data={
-        #     'speclist'              : speclist,
-        #     'pdf'                   : pdf, 
-        #     'x_scale'               : x_scale, 
-        #     'intensities'           : intensities,       
-        #     'fit_results'           : fit_results, 
-        #     'x_fit'                 : x_fit, 
-        #     'd_id'                  : d_id, 
-        #     'scaling_factor'        : scaling_factor, 
-        #     'id_spectra'            : id_spectra
-
-        # },
-        # threads=threads,
-        # close_button=close_button,
-        # progressbar=progress_bars[0]
-        # ))
-        # root.mainloop()
-    logger.info('Save plot to pdf -- Complete')
 
     
 ###################################
